+++ src/RoadCaptain/CryptographyException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace RoadCaptain
+{
+    public class CryptographyException : Exception
+    {
+        public CryptographyException(Exception inner)
+            : base((string?)inner.Message, (Exception?)inner)
+        {
+        }
+    }
+}